Fix ambiguous calls to C standard functions

+++ code/win32/win_main.cpp
@@ -92,9 +92,9 @@ qboolean Sys_FileOutOfDate( LPCSTR psFinalFileName /* dest */, LPCSTR psDataFile
 		// timer res only accurate to within 2 seconds on FAT, so can't do exact compare...
 		//
 		//LONG l = CompareFileTime( &ftFinalFile, &ftDataFile );
-		if (  (abs(ftFinalFile.dwLowDateTime - ftDataFile.dwLowDateTime) <= 20000000 ) &&
-				  ftFinalFile.dwHighDateTime == ftDataFile.dwHighDateTime				
-			)
+		if (abs(static_cast<long long>(ftFinalFile.dwLowDateTime) -
+		        static_cast<long long>(ftDataFile.dwLowDateTime)) <= 20000000 &&
+			ftFinalFile.dwHighDateTime == ftDataFile.dwHighDateTime)
 		{
 			return false;	// file not out of date, ie use it.
 		}
